build(deps): Update soft-fido2 to version 0.3.1
BREAKING CHANGE: The `pass` backend now uses binary credential storage
format instead of JSON. Existing password stores must be recreated as
no migration path is provided. We don't expect more breaking changes
like this in the future. We don't offer a migration mechanism because
probably there are no users but in the future we will offer a way if
this is needed.